refactor: add update login on completing ad payment
