+++ src/main/java/com/myce/advertisement/service/SystemAdService.java
@@ -16,5 +16,5 @@ public interface SystemAdService {
 
     int closeCompletedAds();
 
-    void refreshAdCache();
+    void updateAdStatus();
 }
